backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
dateparser
pytz